Additional condition for filtering TripUpdates
Inspecting the stream, I've found multiple trip updates with the same
trip_id, but different content. I couldn't reliably figure out if that
is a parsing problem, or correct behaviour. So this patch can either be
seen as a work around or a fix for wrong filtering.

If multiple TripUpdates exist for the same trip_id, the last TripUpdate
wins which might not provide stop updates.