Implement schedule query for multiple origin stops
Update 'runSchedule' to query schedules for all the origin stops
given on command line.  Results are merged and sorted via the new
function 'sortSchedules'.

Because the walk delay can differ among origin stops, schedules (and
later in processing, individual schedule items) are now carried
around paired with an associated walk delay.  The 'printSchedule'
type signature was changed to reflect this.

Fixes: https://github.com/romanofski/gtfsschedule/issues/1