Print a message if no services are available
Print a message to let the user know no services are leaving in the next
30 minutes.

This patch has the slight disadvantage because if real time updates are
enabled, they are still being downloaded. Reason why this hasn't changed
is that there is a chance that the real time feed provides information
with added services for a stop. However, displaying added services is
not supported yet.

Bug: #15