Searching for a stop code
This adds a `search` subcommand followed by a substring. The database
query is a like query finding anything with that substring.

Note: Not added additional tests since this is a thin wrapper around a
query. Feels wrong, but I'll go with it at this point and see how much
I'll be shooting myself in the foot.

Issue: #23