Only update vehicle information if stop matches
The vehicle information should only be updated if the stop_id matches
the one stop we're monitoring. Otherwise all vehicle positions will
overwrite vehicle information with information which is unrelated to the
stop we're monitoring.

In case no stop_id is provided, the vehicle information is not updated.
The occupancy status or congestion information is related to a whole
trip at this point which could span over kilometers and who knows how
many stops.