Removing wrong application of unique
The arbitraryUniques1 used nub to remove any additional duplicates in
the generated value. Yet exactly this made it way easier to generate the
same trip_id which is supposed to be unique for all trips. This resulted
in a the likely hood the test to be easily falsifiable, because during
update the schedule items with the same trip_id where assigned to the
same key in the Map used during update.

Also see issue #31