Removing update property test again
This is not a good test, since:

We use a map to lookup schedule items by tripid. They need to be unique
among the generated items. Since QC is stateless, we have no control
over the fact if we haven't generated the same ID already. I could throw
more randomness into it, but at the end it will be a brittle test
because of that.

Closes #31